add null-handling
current implementation fails on initialization of OS enum for any other
platform but WIN, because the ALLUSERSPROFILE is absent. To be sure, we
want to detect that failure on WIN, but it's irrelevant for all other
platforms. By returning a non-null result, we avoid the NPE that occurs
when Paths.get() is invoked, allowing the test to proceed on all other
platforms.

The test still fails on WIN because the expected value is not produced