Added FileFetcher Interface
FileFetcher is a standard interface for fetching resources via I/O.
Implementations of FileFetcher should obtain all the information they
need to fetch a file in their constructor, and then invoke the required
`fetch()` method, which must return an InputStream object.

this allows the interface for fetching from a local file path, a URI, or
a remote resource to be standardized, and eliminates some duplication of
logic.

Also provided the first implementation, which fetches a classpath
resource by its resource name.