Implement DownstreamParamters (Subclass of javafx.application.Application.Parameters)
DownstreamParameters is the catch-all for all arguments passed _to_
FxLauncher, but not used _by_ FxLauncher. FxLauncher assumes these are
intended for the downstream application, and stores them in an instance
of this class.

Subclassing Application.Parameters ensures these options can be passed
to a downstream JavaFX application easily, and providing a getArgs()
method that returns a String array of the rawParams list provides a
means of passing args to non-JavaFX downstream applications